+++ lib/src/core/generators/module_generator.dart
@@ -5,7 +5,7 @@ import 'package:genuis/src/config/types/token_class_type.dart';
 import 'package:genuis/src/core/models/module.dart';
 import 'package:genuis/src/core/writers/module/color_extension_writer.dart';
 import 'package:genuis/src/core/writers/token/enum_token_writer.dart';
-import 'package:genuis/src/core/writers/module/getter_module_writer.dart';
+import 'package:genuis/src/core/writers/module/interface_module_writer.dart';
 import 'package:genuis/src/core/writers/module/theme_extension_module_writer.dart';
 import 'package:genuis/src/core/generators/file_generator.dart';
 import 'package:genuis/src/core/writers/token/static_token_writer.dart';
@@ -79,8 +79,8 @@ class ModuleGenerator extends FileGenerator {
     switch (Config.it.classType) {
       case GenuisClassType.themeExtension:
         const ThemeExtensionModuleWriter().write(buffer, module.rootClass);
-      case GenuisClassType.getter:
-        const GetterModuleWriter().write(buffer, module.rootClass);
+      case GenuisClassType.interface_:
+        const InterfaceModuleWriter().write(buffer, module.rootClass);
     }
 
     return buffer.toString();